requirements.txt
pyyaml
strategoutil